refactor(config-ui): rename StartFormSelector to start-form-selector
